# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "EVHarbor API",
        "termsOfService": "http://obs.casearth.cn/",
        "contact": {
            "name": "API Support",
            "url": "http://obs.casearth.cn/",
            "email": "helpdesk@cnic.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:9000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/obj/{bucketname}/{objpath}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "通过文件对象绝对路径,下载文件对象,可通过参数获取文件对象详细信息，或者自定义读取对象数据块",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "object对象"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bucketname",
                        "name": "bucketname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "objpath",
                        "name": "objpath",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The byte offset of object to read",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Byte size to read",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.UserListJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传对象分片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "object对象"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bucketname",
                        "name": "bucketname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "objpath",
                        "name": "objpath",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "chunk",
                        "name": "chunk",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "chunk_offset",
                        "name": "chunk_offset",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "chunk_size",
                        "name": "chunk_size",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            }
        },
        "/api/v1/users/": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户列表页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user 用户"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The initial index from which to return the results",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to return per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.UserListJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user 用户"
                ],
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.UserProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            }
        },
        "/user/register/": {
            "post": {
                "description": "注册用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register 注册"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Some ID",
                        "name": "some_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The initial index from which to return the results",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to return per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controllers.BaseJSON"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.BaseJSON": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "code_text": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UserListJSON": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "code_text": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "privous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserProfile"
                    }
                }
            }
        },
...
        "models.UserProfile": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}